Fixing duplicate id issue for mount.path
It is dangerous to use {{ mount.path }} as a resource id as it may duplicate defined one.

Example:
Detected conflicting IDs, SLS IDs need to be globally unique.
The conflicting ID is '/var/lib/glance/images' and is found in SLS 'base:glance.server' and SLS 'base:nfs.client'

This patch fixes that issue.

Change-Id: Ie05097be021c1580f4b2805a6ab11899d163d909